refactor(persist): extract shared durable-write + quarantine primitives
Introduce `crate::persist`, the single-source home for the on-disk
persistence primitives the durable stores share: atomic durable write
(tmp → fsync → rename → dir fsync), corruption quarantine + rotation
(rename aside as `.corrupt-<ts>`, preserve bytes), and the fixint bincode
readers with an allocation cap. These lived as `pub(super)`/private
helpers in `user_history` (persistence.rs / recovery.rs), where a second
durable store (user_dict, LXUW) could not reach them — the gap that left
user_dict with an unsynced write and the stray-`.tmp` bug LXUD already
fixed. Centralizing them is the prerequisite for hardening user_dict
without a second, drift-prone copy (CLAUDE.md: 同種処理は単一正準形に収束).

Parity-only: the primitives move verbatim (quarantine generalized to
return an outcome enum + an injected `ts`, so it stays clock-free and
`persist` depends on no store's domain types), and the three consumers —
persistence.rs, recovery.rs, and wal.rs — are redirected to the shared
module. `now_epoch` stays in `user_history` (already the crate-wide time
util, 34 call sites) and is injected into `quarantine`, keeping `persist`
a leaf both stores depend on. Behavior is unchanged; the full suite
(incl. the recovery matrix in tests_recovery.rs) is the proof.
